Adding Participants in Rooms and deleting sms

+++ base/views.py
@@ -61,15 +61,16 @@ def home(request):
 def room(request, pk):
     room = Room.objects.get(id=pk)
     room_messages = room.message_set.all().order_by('-created')
-
+    participants = room.participants.all()
     if request.method == "POST":
         message = Message.objects.create(
             user = request.user,
             room = room,
             body = request.POST.get('body')
         )
+        room.participants.add(request.user)
         return redirect('room', pk= room.id)
-    context = {"room":room, "room_messages": room_messages}
+    context = {"room":room, "room_messages": room_messages, "participants": participants}
     return render(request, 'base/room.html', context)
 
 @login_required(login_url='login')
@@ -101,7 +102,23 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allow here')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj':room})
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('You are not allow here')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj':message})
